refactor: extract interfaces Movie, MovieDetail and TMDBResponse to src/interfaces/tmdb.ts and update all imports.
